// server/src/routes/boards/types.ts
export enum ErrorMessage {
  UNKNOWN_ERROR = 'An unknown error occurred',
  BOARD_NOT_FOUND = 'Board not found',
  COLUMN_NOT_FOUND = 'Column not found',
  INVALID_REQUEST = 'Invalid request data',
  TASK_NOT_FOUND = 'Task not found',
  TASK_CATEGORIES_NOT_FOUND = 'Task categories not found',
  TASK_CATEGORY_NOT_FOUND = 'Task category not found',
  TASK_CATEGORY_ALREADY_EXISTS = 'Task category already exists',
  TASKS_LIST_NOT_FOUND = 'Tasks list not found',
}